fix: Simulation contexts
