+++ src/aux-server/aux-web/aux-player/ContextItem.ts
@@ -1,6 +1,6 @@
 import { Bot } from '@casual-simulation/aux-common';
 
-export interface MenuItem {
+export interface ContextItem {
     bot: Bot;
     simulationId: string;
     contexts: Set<string>;
