+++ src/aux-server/aux-web/aux-player/MenuBot/MenuBot.ts
@@ -8,15 +8,14 @@ import {
     calculateBotValue,
     isFormula,
 } from '@casual-simulation/aux-common';
-// import { MenuItem } from '../MenuContext';
 import { appManager } from '../../shared/AppManager';
-import { MenuItem } from '../MenuItem';
+import { ContextItem } from '../ContextItem';
 
 @Component({
     components: {},
 })
 export default class MenuBot extends Vue {
-    @Prop() item: MenuItem;
+    @Prop() item: ContextItem;
     @Prop() index: number;
     @Prop({ default: false })
     selected: boolean;
@@ -26,7 +25,7 @@ export default class MenuBot extends Vue {
     backgroundColor: string = '#FFF';
 
     @Watch('item')
-    private async _botChanged(item: MenuItem) {
+    private async _botChanged(item: ContextItem) {
         if (item) {
             const simulation = _simulation(item);
             const calc = simulation.helper.createContext();
